some cleanup of the bitcoinj/jtorproxy integration
